+++ libs/i18n/src/lib/translations/translations_es.ts
@@ -1,4 +1,3 @@
-/* eslint-disable */
 // Do not modify, it's automatically created. Modify translations_es.properties instead
 export default {
     coreCalendar: {
@@ -111,9 +110,9 @@ export default {
         collapsedItemMenuLabel: 'Menú de elementos contraídos',
         cancel: 'Cancelar',
         search: 'Buscar',
-        assistiveTools: 'Assistive Tools',
-        backButtonLabel: 'Back',
-        navigationButtonLabel: 'Navigation'
+        assistiveTools: 'Herramientas de asistencia',
+        backButtonLabel: 'Atrás',
+        navigationButtonLabel: 'Navegación'
     },
     coreSlider: {
         singleMinMaxDetails: 'El valor mínimo del slider es {min}, el valor máximo es {max}',
@@ -201,7 +200,7 @@ export default {
         ariaLabel: 'Asistente'
     },
     coreBreadcrumb: {
-        overflowTitleMore: 'Más',
+        overflowTitleMore: 'Hacer clic o pulsar intro para ver más detalles',
         breadcrumbTrailLabel: 'Ruta de navegación'
     },
     platformApprovalFlow: {
@@ -536,7 +535,8 @@ export default {
         confirmBtnLabel: 'Aceptar',
         cancelBtnLabel: 'Cancelar',
         filteredBy: 'Filtrado por',
-        showingBlankOfBlank: 'Mostrando {totalGrowingItemsCount} de {totalGrowingItemsCount}'
+        showingBlankOfBlank: 'Mostrando {totalGrowingItemsCount} de {totalGrowingItemsCount}',
+        clearFilters: 'Borrar filtros'
     },
     platformWizardGenerator: {
         summarySectionEditStep: 'Editar'
@@ -601,7 +601,7 @@ export default {
         groupAriaDescription: 'Grupo de notificaciones',
         groupAriaDescriptionExpanded: 'expandido',
         groupAriaDescriptionCollapsed: 'contraído',
-        triggerMoreLabel: 'More',
-        triggerLessLabel: 'Less'
+        triggerMoreLabel: 'Más',
+        triggerLessLabel: 'Menos'
     }
 };